+++ sentry_sdk/sessions.py
@@ -1,27 +1,21 @@
 import os
-import uuid
 import time
-from datetime import datetime
 from threading import Thread, Lock
 from contextlib import contextmanager
 
 import sentry_sdk
+from sentry_sdk.envelope import Envelope
+from sentry_sdk.session import Session
 from sentry_sdk._types import MYPY
 from sentry_sdk.utils import format_timestamp
 
 if MYPY:
+    from typing import Callable
     from typing import Optional
-    from typing import Union
     from typing import Any
     from typing import Dict
     from typing import List
     from typing import Generator
-
-    from sentry_sdk._types import SessionStatus
-
-
-def minute_trunc(ts):
-    return ts.replace(second=0, microsecond=0)
 
 
 def is_auto_session_tracking_enabled(hub=None):
@@ -52,30 +46,19 @@ def auto_session_tracking(hub=None):
             hub.end_session()
 
 
-def _make_uuid(
-    val,  # type: Union[str, uuid.UUID]
-):
-    # type: (...) -> uuid.UUID
-    if isinstance(val, uuid.UUID):
-        return val
-    return uuid.UUID(val)
-
-
 TERMINAL_SESSION_STATES = ("exited", "abnormal", "crashed")
 MAX_ENVELOPE_ITEMS = 100
 
 
 def make_aggregate_envelope(aggregate_states, attrs):
-    rv = {"attrs": dict(attrs), "aggregates": []}
-    for state in aggregate_states.values():
-        rv["aggregates"].append(state)
-    return rv
+    # type: (Any, Any) -> Any
+    return {"attrs": dict(attrs), "aggregates": list(aggregate_states.values())}
 
 
 class SessionFlusher(object):
     def __init__(
         self,
-        capture_func,  # type: (sentry_sdk.envelope.Envelope) -> None
+        capture_func,  # type: Callable[[Envelope], None]
         session_mode,  # type: str
         flush_interval=60,  # type: int
     ):
@@ -99,9 +82,6 @@ class SessionFlusher(object):
         with self._aggregate_lock:
             pending_aggregates = self.pending_aggregates
             self.pending_aggregates = {}
-
-        # NOTE: use absolute import here to avoid circular imports
-        Envelope = sentry_sdk.envelope.Envelope  # noqa
 
         envelope = Envelope()
         for session in pending_sessions:
@@ -146,6 +126,7 @@ class SessionFlusher(object):
     def add_aggregate_session(
         self, session  # type: Session
     ):
+        # type: (...) -> None
         # NOTE on `session.did`:
         # the protocol can deal with buckets that have a distinct-id, however
         # in practice we expect the python SDK to have an extremely high cardinality
@@ -190,145 +171,3 @@ class SessionFlusher(object):
     def __del__(self):
         # type: (...) -> None
         self.kill()
-
-
-class Session(object):
-    def __init__(
-        self,
-        sid=None,  # type: Optional[Union[str, uuid.UUID]]
-        did=None,  # type: Optional[str]
-        timestamp=None,  # type: Optional[datetime]
-        started=None,  # type: Optional[datetime]
-        duration=None,  # type: Optional[float]
-        status=None,  # type: Optional[SessionStatus]
-        release=None,  # type: Optional[str]
-        environment=None,  # type: Optional[str]
-        user_agent=None,  # type: Optional[str]
-        ip_address=None,  # type: Optional[str]
-        errors=None,  # type: Optional[int]
-        user=None,  # type: Optional[Any]
-    ):
-        # type: (...) -> None
-        if sid is None:
-            sid = uuid.uuid4()
-        if started is None:
-            started = datetime.utcnow()
-        if status is None:
-            status = "ok"
-        self.status = status
-        self.did = None  # type: Optional[str]
-        self.started = started
-        self.release = None  # type: Optional[str]
-        self.environment = None  # type: Optional[str]
-        self.duration = None  # type: Optional[float]
-        self.user_agent = None  # type: Optional[str]
-        self.ip_address = None  # type: Optional[str]
-        self.errors = 0
-
-        self.update(
-            sid=sid,
-            did=did,
-            timestamp=timestamp,
-            duration=duration,
-            release=release,
-            environment=environment,
-            user_agent=user_agent,
-            ip_address=ip_address,
-            errors=errors,
-            user=user,
-        )
-
-    @property
-    def truncated_started(self):
-        return minute_trunc(self.started)
-
-    def update(
-        self,
-        sid=None,  # type: Optional[Union[str, uuid.UUID]]
-        did=None,  # type: Optional[str]
-        timestamp=None,  # type: Optional[datetime]
-        started=None,  # type: Optional[datetime]
-        duration=None,  # type: Optional[float]
-        status=None,  # type: Optional[SessionStatus]
-        release=None,  # type: Optional[str]
-        environment=None,  # type: Optional[str]
-        user_agent=None,  # type: Optional[str]
-        ip_address=None,  # type: Optional[str]
-        errors=None,  # type: Optional[int]
-        user=None,  # type: Optional[Any]
-    ):
-        # type: (...) -> None
-        # If a user is supplied we pull some data form it
-        if user:
-            if ip_address is None:
-                ip_address = user.get("ip_address")
-            if did is None:
-                did = user.get("id") or user.get("email") or user.get("username")
-
-        if sid is not None:
-            self.sid = _make_uuid(sid)
-        if did is not None:
-            self.did = str(did)
-        if timestamp is None:
-            timestamp = datetime.utcnow()
-        self.timestamp = timestamp
-        if started is not None:
-            self.started = started
-        if duration is not None:
-            self.duration = duration
-        if release is not None:
-            self.release = release
-        if environment is not None:
-            self.environment = environment
-        if ip_address is not None:
-            self.ip_address = ip_address
-        if user_agent is not None:
-            self.user_agent = user_agent
-        if errors is not None:
-            self.errors = errors
-
-        if status is not None:
-            self.status = status
-
-    def close(
-        self, status=None  # type: Optional[SessionStatus]
-    ):
-        # type: (...) -> Any
-        if status is None and self.status == "ok":
-            status = "exited"
-        if status is not None:
-            self.update(status=status)
-
-    def get_json_attrs(self, with_user_info=True):
-        # type: (...) -> Any
-        attrs = {}
-        if self.release is not None:
-            attrs["release"] = self.release
-        if self.environment is not None:
-            attrs["environment"] = self.environment
-        if with_user_info:
-            if self.ip_address is not None:
-                attrs["ip_address"] = self.ip_address
-            if self.user_agent is not None:
-                attrs["user_agent"] = self.user_agent
-        return attrs
-
-    def to_json(self):
-        # type: (...) -> Any
-        rv = {
-            "sid": str(self.sid),
-            "init": True,
-            "started": format_timestamp(self.started),
-            "timestamp": format_timestamp(self.timestamp),
-            "status": self.status,
-        }  # type: Dict[str, Any]
-        if self.errors:
-            rv["errors"] = self.errors
-        if self.did is not None:
-            rv["did"] = self.did
-        if self.duration is not None:
-            rv["duration"] = self.duration
-        attrs = self.get_json_attrs()
-        if attrs:
-            rv["attrs"] = attrs
-        return rv
